python/requirements.txt
pillow>=10.0.0
rembg>=2.0.50
onnxruntime>=1.16.0
numpy>=1.24.0
